requirements.txt
requests~=2.31.0
websocket-client~=1.6.2